- Implemented the interfaces

+++ src/main/java/com/remmelt/example/repository/PersonRepository.java
@@ -3,7 +3,7 @@ package com.remmelt.example.repository;
 import com.remmelt.example.model.Person;
 
 public interface PersonRepository {
-	Person getPersonBy(Long id) ;
+	Person getPersonBy(int id) ;
 
 	void savePerson(Person person);
 }
